+++ hypervisor/include/arch/x86/irq.h
@@ -119,7 +119,7 @@ void interrupt_init(uint16_t pcpu_id);
 void cancel_event_injection(struct vcpu *vcpu);
 
 #ifdef HV_DEBUG
-void get_cpu_interrupt_info(char *str_arg, int str_max);
+void get_cpu_interrupt_info(char *str_arg, size_t str_max);
 #endif /* HV_DEBUG */
 
 extern uint32_t acrn_vhm_vector;
